Moved voice note logic to audioRecording class. Added recording microphone image to voice note button
