+++ web/static_assets/index.html
@@ -15,16 +15,13 @@
 
           <div class="col-4">
             <button class="button" id="new-note">
-                <div><img src="images/plus-sign.png"></div>
+                <div><img src="images/plus-sign.png"> New Note</div>
             </button>
-            <button class="button" id="start-recording">
-              <div><img src="images/plus-sign.png"> Start</div>
+            <button class="button" id="new-voice-note">
+              <div><img src="images/recording.png"> Voice Note</div>
             </button>
             <button class="button" id="stop-recording">
-              <div><img src="images/plus-sign.png"> Stop</div>
-            </button>
-            <button class="button" id="voice-note-test">
-              <div><img src="images/plus-sign.png"> Test</div>
+              <div><img src="images/plus-sign.png"> Voice Note</div>
             </button>
             <div class="note-sort-and-search">
                 <button class="dropdown-toggle" type="button" data-bs-toggle="dropdown" id="sort-by">Sort By</button>
